+++ apps/www/src/app/(main)/(docs)/components/data-table/examples/usage.tsx
@@ -3,9 +3,9 @@
 import { DataTable, type DataTableProps } from "@jamsr-ui/react";
 import { SortingState } from "@tanstack/react-table";
 import { useState } from "react";
-import { COLUMNS, USERS } from "./columns";
+import { COLUMNS, User, USERS } from "./columns";
 
-export const DataTableUsage = (props: Partial<DataTableProps>) => {
+export const DataTableUsage = (props: Partial<DataTableProps<User>>) => {
   const [sorting, setSorting] = useState<SortingState>([
     { desc: true, id: "registeredAt" },
   ]);
@@ -14,7 +14,6 @@ export const DataTableUsage = (props: Partial<DataTableProps>) => {
     <DataTable
       columns={COLUMNS}
       data={USERS}
-      rowCount={USERS.length}
       onSortingChange={setSorting}
       state={{
         sorting,