+++ packages/components/data-table/src/types.d.ts
@@ -2,6 +2,6 @@ import { type DataTableProps } from ".";
 
 declare module "@jamsr-ui/config" {
   export interface UIConfigType {
-    dataTable?: Partial<DataTableProps>;
+    dataTable?: Partial<DataTableProps<any>>;
   }
 }