+++ packages/components/data-table/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @jamsr-ui/data-table
 
+## 0.1.9
+
+### Patch Changes
+
+- fix: data table types
+
 ## 0.1.8
 
 ### Patch Changes